# openapi/api.yaml
openapi: 3.1.0
info:
  title: Karelia API Programming - OpenAPI 3.1.0 Demonstration
  description: |-
    This is a sample OpenAPI OAD.
    
    Some useful links:
    - [Learn OpenAPI](https://learn.openapis.org)
    - [OpenAPI Specification](https://learn.openapis.org/)
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.1.1

# 2) Apply the security globally to all operations
security:
  - bearerAuth: []

servers:
  - url: http://localhost:3000
  - url: anatoliiu2306851.westeurope.cloudapp.azure.com
  
tags:
  - name: Sample dataset
    description: Everything you need to know about APIs
    externalDocs:
      description: Find out more
      url: https://m.karelia.fi

paths:
  /data:
    get:
      tags:
        - Sample dataset
      summary: Returns a dataset
      description: Returns a dataset with something more blaah...
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Data'
        '415':
          description: Unsupported media type
  /data/insert-rows:
      post:
        tags:
          - Sample dataset
        summary: Inserts multiple rows of data
        description: Inserts a specified number of rows into the dataset.
        requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  num_rows:
                    type: integer
                    description: "The number of rows to insert."
        responses:
          '200':
            description: Rows inserted successfully
          '400':
            description: Invalid number of rows
          '500':
            description: Failed to insert rows

components:
  schemas:
    Data:
      type: object
      properties:
        id:
          type: integer
          format: int64
          examples:
            - 111
        forename:
          type: string
          examples:
            - Roy
          
        surname:
          type: string
          examples: [ Fielding ]

  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT